refactor: 연동 작업 조회 API 스펙 개선 및 리포지토리 구조화 (#64)
* refactor: SyncJobQueryCondition으로 네이밍 수정

* refactor: SyncJobQueryCondition 시간 타입 변경(LocalDate) 및 Repository 변환 로직 추가

* refactor: QueryDSL 리포지토리 패키지 구조 및 네이밍 컨벤션 적용

* docs: Swagger 연동 작업 API 명세서 텍스트 통일

* fix: 작업 종료일 검색 시 서브초(sub-second) 단위 데이터 누락 버그 수정
