+++ src/main/java/com/sprint/mission/findex/domain/syncjob/controller/SyncJobController.java
@@ -4,7 +4,7 @@ import com.sprint.mission.findex.domain.syncjob.controller.api.SyncJobApi;
 import com.sprint.mission.findex.domain.syncjob.dto.IndexDataSyncRequest;
 import com.sprint.mission.findex.domain.syncjob.dto.IndexInfoSyncRequest;
 import com.sprint.mission.findex.domain.syncjob.dto.SyncJobResponse;
-import com.sprint.mission.findex.domain.syncjob.dto.SyncJobSearchCondition;
+import com.sprint.mission.findex.domain.syncjob.dto.SyncJobQueryCondition;
 import com.sprint.mission.findex.domain.syncjob.service.SyncJobService;
 import com.sprint.mission.findex.global.common.dto.CursorPageResponse;
 import jakarta.servlet.http.HttpServletRequest;
@@ -55,7 +55,7 @@ public class SyncJobController implements SyncJobApi {
   @GetMapping
   @Override
   public ResponseEntity<CursorPageResponse<SyncJobResponse>> getSyncJobHistory(
-      @Valid @ParameterObject @ModelAttribute SyncJobSearchCondition condition,
+      @Valid @ParameterObject @ModelAttribute SyncJobQueryCondition condition,
       @RequestParam(required = false) String cursor,
       @RequestParam(required = false) UUID idAfter,
       @RequestParam(defaultValue = "jobTime") String sortField,
