+++ src/main/java/com/sprint/mission/findex/domain/syncjob/controller/api/SyncJobApi.java
@@ -3,7 +3,7 @@ package com.sprint.mission.findex.domain.syncjob.controller.api;
 import com.sprint.mission.findex.domain.syncjob.dto.IndexDataSyncRequest;
 import com.sprint.mission.findex.domain.syncjob.dto.IndexInfoSyncRequest;
 import com.sprint.mission.findex.domain.syncjob.dto.SyncJobResponse;
-import com.sprint.mission.findex.domain.syncjob.dto.SyncJobSearchCondition;
+import com.sprint.mission.findex.domain.syncjob.dto.SyncJobQueryCondition;
 import com.sprint.mission.findex.global.common.dto.CursorPageResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -19,7 +19,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import java.util.List;
 import java.util.UUID;
 
-@Tag(name = "Sync Job", description = "지수 데이터 연동 및 이력 관리 API (M4)")
+@Tag(name = "연동 작업 API", description = "연동 작업 관리 API")
 public interface SyncJobApi {
 
   @Operation(summary = "지수 정보 연동", description = "KRX Open API에서 기준 날짜의 전체 지수 정보를 조회하여 DB에 생성 또는 갱신하고 SyncJob 이력을 저장합니다.")
@@ -27,7 +27,7 @@ public interface SyncJobApi {
       @Valid @RequestBody IndexInfoSyncRequest request,
       HttpServletRequest servletRequest);
 
-  @Operation(summary = "지수 데이터 수동 연동", description = "지정한 기간(baseDateFrom ~ baseDateTo)의 하나 이상의 지수 데이터를 외부 API로부터 연동합니다.")
+  @Operation(summary = "지수 데이터 연동", description = "지정한 기간(baseDateFrom ~ baseDateTo)의 하나 이상의 지수 데이터를 외부 API로부터 연동합니다.")
   ResponseEntity<List<SyncJobResponse>> syncIndexData(
       @Valid @RequestBody IndexDataSyncRequest request,
       HttpServletRequest servletRequest);
@@ -35,7 +35,7 @@ public interface SyncJobApi {
   @Operation(summary = "연동 작업 목록 조회", description = "정렬값(cursor)과 ID(idAfter) 기반 복합 커서를 사용하여 연동 이력을 페이징 조회합니다.")
   ResponseEntity<CursorPageResponse<SyncJobResponse>> getSyncJobHistory(
 
-      @Valid @ParameterObject @ModelAttribute SyncJobSearchCondition condition,
+      @Valid @ParameterObject @ModelAttribute SyncJobQueryCondition condition,
 
       @Parameter(description = "커서 (다음 페이지 시작점)")
       @RequestParam(required = false) String cursor,
