+++ src/main/java/com/sprint/mission/findex/domain/syncjob/repository/querydsl/SyncJobCustomRepository.java
@@ -1,7 +1,7 @@
-package com.sprint.mission.findex.domain.syncjob.repository;
+package com.sprint.mission.findex.domain.syncjob.repository.querydsl;
 
 import com.sprint.mission.findex.domain.syncjob.dto.SyncJobResponse;
-import com.sprint.mission.findex.domain.syncjob.dto.SyncJobSearchCondition;
+import com.sprint.mission.findex.domain.syncjob.dto.SyncJobQueryCondition;
 import com.sprint.mission.findex.global.common.dto.CursorPageResponse;
 
 import java.util.UUID;
@@ -9,7 +9,7 @@ import java.util.UUID;
 public interface SyncJobCustomRepository {
 
   CursorPageResponse<SyncJobResponse> searchSyncJobPage(
-      SyncJobSearchCondition condition,
+      SyncJobQueryCondition condition,
       String cursor,
       UUID idAfter,
       String sortField,
