+++ src/main/java/com/sprint/mission/findex/domain/syncjob/service/SyncJobService.java
@@ -10,7 +10,7 @@ import com.sprint.mission.findex.domain.indexinfo.repository.IndexInfoRepository
 import com.sprint.mission.findex.domain.syncclient.client.KrxOpenApiClient;
 import com.sprint.mission.findex.domain.syncclient.dto.IndexDataApiResponse;
 import com.sprint.mission.findex.domain.syncjob.dto.SyncJobResponse;
-import com.sprint.mission.findex.domain.syncjob.dto.SyncJobSearchCondition;
+import com.sprint.mission.findex.domain.syncjob.dto.SyncJobQueryCondition;
 import com.sprint.mission.findex.domain.syncjob.entity.JobResult;
 import com.sprint.mission.findex.domain.syncjob.entity.JobType;
 import com.sprint.mission.findex.domain.syncjob.entity.SyncJob;
@@ -132,7 +132,7 @@ public class SyncJobService {
 
   @Transactional(readOnly = true)
   public CursorPageResponse<SyncJobResponse> getSyncJobHistory(
-      SyncJobSearchCondition condition,
+      SyncJobQueryCondition condition,
       String cursor,
       UUID idAfter,
       String sortField,
